# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/requests": {
      "post": {
        "description": "Réservé à un hôpital dont l'établissement est validé. Si le score d'anomalie est élevé, la demande passe en en_revue et aucune alerte n'est envoyée avant validation manuelle. Sinon la vague 1 démarre immédiatement (moins de 10 s).",
        "operationId": "RequestsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Token absent, invalide ou expiré",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Rôle insuffisant. Rôles autorisés : hopital",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "422": {
            "description": "Demande refusée par la détection d'anomalies (échéance déjà dépassée)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Créer une demande urgente",
        "tags": [
          "requests"
        ],
        "x-roles": [
          "hopital"
        ]
      },
      "get": {
        "description": "Pour un hôpital : ses demandes. Pour la direction et l'admin : toutes.",
        "operationId": "RequestsController_list",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/RequestStatus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RequestDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token absent, invalide ou expiré",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Rôle insuffisant. Rôles autorisés : hopital, direction, admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lister les demandes",
        "tags": [
          "requests"
        ],
        "x-roles": [
          "hopital",
          "direction",
          "admin"
        ]
      }
    },
    "/requests/review": {
      "get": {
        "description": "Demandes en en_revue, score d'anomalie décroissant. Aucune alerte n'est partie pour elles tant qu'un admin n'a pas décidé.",
        "operationId": "RequestsController_listForReview",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RequestDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token absent, invalide ou expiré",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Rôle insuffisant. Rôles autorisés : admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Demandes retenues en revue (admin)",
        "tags": [
          "requests"
        ],
        "x-roles": [
          "admin"
        ]
      }
    },
    "/requests/{id}": {
      "get": {
        "operationId": "RequestsController_getOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": "b2c3d4e5-0001-4bbb-9ccc-000000000001",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestDto"
                }
              }
            }
          },
          "401": {
            "description": "Token absent, invalide ou expiré",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Rôle insuffisant. Rôles autorisés : hopital, direction, admin, donneur",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Détail d'une demande",
        "tags": [
          "requests"
        ],
        "x-roles": [
          "hopital",
          "direction",
          "admin",
          "donneur"
        ]
      }
    },
...
    "/requests/{id}/review": {
      "patch": {
        "description": "approve : la demande passe en active et la vague 1 démarre. reject : la demande est clôturée sans aucune alerte.",
        "operationId": "RequestsController_review",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": "b2c3d4e5-0001-4bbb-9ccc-000000000001",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestDto"
                }
              }
            }
          },
          "401": {
            "description": "Token absent, invalide ou expiré",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Rôle insuffisant. Rôles autorisés : admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "La demande n'est pas en revue, ou son échéance est dépassée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Valider ou rejeter une demande en revue (admin)",
        "tags": [
          "requests"
        ],
        "x-roles": [
          "admin"
        ]
      }
    },
    "/requests/{id}/respond": {
      "post": {
        "description": "Contrôle serveur : demande encore active, donneur compatible et éligible, délai entre dons écoulé (3ᵉ niveau de F2.4). Une seule réponse par donneur et par demande.",
        "operationId": "RequestsController_respond",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": "b2c3d4e5-0001-4bbb-9ccc-000000000001",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RespondDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespondResultDto"
                }
              }
            }
          },
          "401": {
            "description": "Token absent, invalide ou expiré",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Rôle insuffisant. Rôles autorisés : donneur",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Déjà répondu, demande close, ou donneur non éligible à ce jour",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Répondre à une alerte (Je viens / Je ne peux pas)",
        "tags": [
          "requests"
        ],
        "x-roles": [
          "donneur"
        ]
      }
    },
...
  },
  "info": {
    "title": "Damm API",
    "description": "Plateforme intelligente de don de sang (HackCure 48 h).\n\nContrat v1 : toute modification passe par une PR sur docs/openapi.json et est annoncée à l'équipe.\nLe suivi en direct (WebSocket /live) est décrit dans docs/ws-live.md ; ses charges utiles\net celle de l'alerte FCM figurent dans les schémas WsGaugeEventDto, WsDonorEnRouteEventDto,\nWsWaveStartedEventDto et AlertPayloadDto.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "message": {
            "description": "Message lisible, ou liste de messages de validation",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": [
              "phone must be a valid phone number"
            ]
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
...
      "BloodGroup": {
        "type": "string",
        "enum": [
          "A+",
          "A-",
          "B+",
          "B-",
          "AB+",
          "AB-",
          "O+",
          "O-"
        ],
        "description": "Groupe déclaré, à confirmer au 1er don"
      },
...
      "UrgencyLevel": {
        "type": "string",
        "enum": [
          "normale",
          "urgente",
          "critique"
        ]
      },
      "CreateRequestDto": {
        "type": "object",
        "properties": {
          "bloodGroup": {
            "example": "A+",
            "allOf": [
              {
                "$ref": "#/components/schemas/BloodGroup"
              }
            ]
          },
          "quantity": {
            "type": "number",
            "example": 10,
            "minimum": 1,
            "description": "Nombre de poches demandées"
          },
          "urgency": {
            "example": "urgente",
            "allOf": [
              {
                "$ref": "#/components/schemas/UrgencyLevel"
              }
            ]
          },
          "deadline": {
            "type": "string",
            "example": "2026-10-03T18:00:00.000Z",
            "description": "Échéance au-delà de laquelle la demande expire"
          },
          "initialRadiusKm": {
            "type": "number",
            "enum": [
              5,
              10,
              20
            ],
            "default": 10,
            "description": "Rayon initial en km, choisi par l'hôpital (F3.2)"
          }
        },
        "required": [
          "bloodGroup",
          "quantity",
          "urgency",
          "deadline"
        ]
      },
      "RequestStatus": {
        "type": "string",
        "enum": [
          "en_revue",
          "active",
          "couverte",
          "cloturee",
          "expiree"
        ],
        "description": "en_revue = score d'anomalie élevé, validation manuelle requise avant toute alerte"
      },
      "RequestDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "b2c3d4e5-0001-4bbb-9ccc-000000000001"
          },
          "institutionId": {
            "type": "string",
            "example": "a1b2c3d4-0001-4aaa-8bbb-000000000001"
          },
          "institutionName": {
            "type": "string",
            "example": "Hôpital Charles Nicolle"
          },
          "bloodGroup": {
            "allOf": [
              {
                "$ref": "#/components/schemas/BloodGroup"
              }
            ]
          },
          "quantity": {
            "type": "number",
            "example": 10
          },
          "urgency": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UrgencyLevel"
              }
            ]
          },
          "deadline": {
            "type": "string",
            "example": "2026-10-03T18:00:00.000Z"
          },
          "initialRadiusKm": {
            "type": "number",
            "example": 5
          },
          "currentRadiusKm": {
            "type": "number",
            "example": 5,
            "description": "Rayon actuel, élargi à chaque vague"
          },
          "maxRadiusKm": {
            "type": "number",
            "example": 30
          },
          "status": {
            "description": "en_revue = score d'anomalie élevé, validation manuelle requise avant toute alerte",
            "allOf": [
              {
                "$ref": "#/components/schemas/RequestStatus"
              }
            ]
          },
          "anomalyScore": {
            "type": "number",
            "example": 0.12,
            "description": "Score d'anomalie de 0 à 1 (M2)"
          },
          "createdAt": {
            "type": "string",
            "example": "2026-10-03T10:30:00.000Z"
          }
        },
        "required": [
          "id",
          "institutionId",
          "institutionName",
          "bloodGroup",
          "quantity",
          "urgency",
          "deadline",
          "initialRadiusKm",
          "currentRadiusKm",
          "maxRadiusKm",
          "status",
          "createdAt"
        ]
      },
      "GaugeDto": {
        "type": "object",
        "properties": {
          "accepted": {
            "type": "number",
            "example": 3,
            "description": "Donneurs ayant répondu « Je viens »"
          },
          "needed": {
            "type": "number",
            "example": 10,
            "description": "Quantité demandée"
          },
          "percent": {
            "type": "number",
            "example": 30,
            "description": "Pourcentage de couverture, plafonné à 100"
          }
        },
        "required": [
          "accepted",
          "needed",
          "percent"
        ]
      },
...
      "ReviewDecision": {
        "type": "string",
        "enum": [
          "approve",
          "reject"
        ],
        "description": "approve : la demande passe en active et la vague 1 démarre. reject : la demande est clôturée, aucune alerte envoyée."
      },
      "ReviewRequestDto": {
        "type": "object",
        "properties": {
          "decision": {
            "example": "approve",
            "description": "approve : la demande passe en active et la vague 1 démarre. reject : la demande est clôturée, aucune alerte envoyée.",
            "allOf": [
              {
                "$ref": "#/components/schemas/ReviewDecision"
              }
            ]
          }
        },
        "required": [
          "decision"
        ]
      },
      "ResponseType": {
        "type": "string",
        "enum": [
          "je_viens",
          "ne_peut_pas"
        ]
      },
      "RespondDto": {
        "type": "object",
        "properties": {
          "response": {
            "example": "je_viens",
            "allOf": [
              {
                "$ref": "#/components/schemas/ResponseType"
              }
            ]
          }
        },
        "required": [
          "response"
        ]
      },
      "RespondResultDto": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string",
            "example": "b2c3d4e5-0001-4bbb-9ccc-000000000001"
          },
          "response": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ResponseType"
              }
            ]
          },
          "accepted": {
            "type": "boolean",
            "example": true,
            "description": "false si le contrôle serveur refuse (délai entre dons non écoulé, demande close). Voir le code HTTP 409."
          },
          "gauge": {
            "$ref": "#/components/schemas/GaugeDto"
          }
        },
        "required": [
          "requestId",
          "response",
          "accepted",
          "gauge"
        ]
      },
...
    }
  }
}